// lib/schema.ts
export interface Database {
  public: {
    Tables: {
      profiles: {
        Row: {
          biography: string | null
          display_name: string
          email: string
          id: string
        }
        Insert: {
          biography?: string | null
          display_name: string
          email: string
          id: string
        }
        Update: {
          biography?: string | null
          display_name?: string
          email?: string
          id?: string
        }
        Relationships: [
          {
            foreignKeyName: "profiles_id_fkey"
            columns: ["id"]
            isOneToOne: true
            referencedRelation: "users"
            referencedColumns: ["id"]
          }
        ]
      }
      species: {
        Row: {
          author: string
          common_name: string | null
          description: string | null
          id: number
          image: string | null
          kingdom: Database["public"]["Enums"]["kingdom"]
          scientific_name: string
          total_population: number | null
        }
        Insert: {
          author: string
          common_name?: string | null
          description?: string | null
          id?: number
          image?: string | null
          kingdom: Database["public"]["Enums"]["kingdom"]
          scientific_name: string
          total_population?: number | null
        }
        Update: {
          author?: string
          common_name?: string | null
          description?: string | null
          id?: number
          image?: string | null
          kingdom?: Database["public"]["Enums"]["kingdom"]
          scientific_name?: string
          total_population?: number | null
        }
        Relationships: [
          {
            foreignKeyName: "species_author_fkey"
            columns: ["author"]
            isOneToOne: false
            referencedRelation: "profiles"
            referencedColumns: ["id"]
          }
        ]
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      kingdom:
        | "Animalia"
        | "Plantae"
        | "Fungi"
        | "Protista"
        | "Archaea"
        | "Bacteria"
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

export type Tables<
  PublicTableNameOrOptions extends
    | keyof (Database["public"]["Tables"] & Database["public"]["Views"])
    | { schema: keyof Database },
  TableName extends PublicTableNameOrOptions extends { schema: keyof Database }
    ? keyof (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
        Database[PublicTableNameOrOptions["schema"]]["Views"])
    : never = never
> = PublicTableNameOrOptions extends { schema: keyof Database }
  ? (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
      Database[PublicTableNameOrOptions["schema"]]["Views"])[TableName] extends {
      Row: infer R
    }
    ? R
    : never
  : PublicTableNameOrOptions extends keyof (Database["public"]["Tables"] &
      Database["public"]["Views"])
  ? (Database["public"]["Tables"] &
      Database["public"]["Views"])[PublicTableNameOrOptions] extends {
      Row: infer R
    }
    ? R
    : never
  : never
// ...
export type Enums<
  PublicEnumNameOrOptions extends
    | keyof Database["public"]["Enums"]
    | { schema: keyof Database },
  EnumName extends PublicEnumNameOrOptions extends { schema: keyof Database }
    ? keyof Database[PublicEnumNameOrOptions["schema"]]["Enums"]
    : never = never
> = PublicEnumNameOrOptions extends { schema: keyof Database }
  ? Database[PublicEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : PublicEnumNameOrOptions extends keyof Database["public"]["Enums"]
  ? Database["public"]["Enums"][PublicEnumNameOrOptions]
  : never